feat(ui): refactor menu
