+++ ui/src/app/modules/agent-network/agent-network.routes.ts
@@ -1,7 +1,5 @@
 import { AgentNetworkComponent } from './components/agent-network/agent-network.component';
 import { Routes } from '@angular/router';
-import { ChutneyMainHeaderComponent } from '@shared/components/chutney-main-header/chutney-main-header.component';
-import { ChutneyLeftMenuComponent } from '@shared/components/chutney-left-menu/chutney-left-menu.component';
 
 export const AgentNetworkRoute: Routes = [
     {
