+++ ui/src/app/modules/global-variable/global-variable.routes.ts
@@ -3,8 +3,6 @@ import { GlobalVariableEditionComponent } from './components/global-variable-edi
 
 import { AuthGuard } from '@core/guards';
 import { Authorization } from '@model';
-import { ChutneyLeftMenuComponent } from '@shared/components/chutney-left-menu/chutney-left-menu.component';
-import { ChutneyMainHeaderComponent } from '@shared/components/chutney-main-header/chutney-main-header.component';
 
 export const GlobalVariableRoute: Routes = [
     {
