+++ ui/src/app/modules/roles/roles.routes.ts
@@ -1,7 +1,5 @@
 import { Routes } from '@angular/router';
 import { RolesComponent } from './components/roles.component';
-import { ChutneyMainHeaderComponent } from '@shared/components/chutney-main-header/chutney-main-header.component';
-import { ChutneyLeftMenuComponent } from '@shared/components/chutney-left-menu/chutney-left-menu.component';
 
 export const RolesRoute: Routes = [
     {
