+++ ui/src/app/organisms/environment-admin/environment-admin.routes.ts
@@ -3,8 +3,6 @@ import { Routes } from '@angular/router';
 import { EnvironmentAdminComponent } from './environment-admin.component';
 import { AuthGuard } from '@core/guards';
 import { Authorization } from '@model';
-import { ChutneyMainHeaderComponent } from '@shared/components/chutney-main-header/chutney-main-header.component';
-import { ChutneyLeftMenuComponent } from '@shared/components/chutney-left-menu/chutney-left-menu.component';
 
 export const environmentAdminRoute: Routes = [
     {
