+++ ui/src/app/shared/components/layout/header/chutney-main-header.component.scss
