+++ ui/src/app/shared/components/layout/left-menu/chutney-left-menu.component.html
