+++ ui/src/app/shared/components/layout/right-menu/chutney-right-menu.component.scss
