+++ ui/src/app/shared/shared.module.ts
@@ -36,11 +36,12 @@ import {
 import { StateService } from './state/state.service';
 import { ChutneyEditorComponent } from '@shared/components/chutney-editor/chutney-editor.component';
 import { HjsonParserService } from '@shared/hjson-parser/hjson-parser.service';
-import { ChutneyMainHeaderComponent } from './components/chutney-main-header/chutney-main-header.component';
-import { ChutneyLeftMenuComponent } from './components/chutney-left-menu/chutney-left-menu.component';
-import { ChutneyRightMenuComponent } from './components/chutney-right-menu/chutney-right-menu.component';
+
 import { PerfectScrollbarModule } from 'ngx-perfect-scrollbar';
 import { NgbDropdownModule, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
+import { ChutneyMainHeaderComponent } from '@shared/components/layout/header/chutney-main-header.component';
+import { ChutneyLeftMenuComponent } from '@shared/components/layout/left-menu/chutney-left-menu.component';
+import { ChutneyRightMenuComponent } from '@shared/components/layout/right-menu/chutney-right-menu.component';
 
 @NgModule({
     imports: [
